[hotifx] Relocate Action class.
fix Action location.
