+++ plan/src/main/java/org/apache/flink/agents/plan/serializer/AgentPlanJsonDeserializer.java
@@ -19,9 +19,9 @@
 package org.apache.flink.agents.plan.serializer;
 
 import org.apache.flink.agents.api.resource.ResourceType;
-import org.apache.flink.agents.plan.Action;
 import org.apache.flink.agents.plan.AgentConfiguration;
 import org.apache.flink.agents.plan.AgentPlan;
+import org.apache.flink.agents.plan.actions.Action;
 import org.apache.flink.agents.plan.resourceprovider.ResourceProvider;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.core.JacksonException;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.core.JsonParser;
